+++ src/index.js
@@ -1,0 +1,3 @@
+import { WysiwygController, detectUnsupportedMarkdown, parseMarkdown, serializeMarkdown } from './wysiwyg-controller.js';
+
+window.KlockWysiwyg = { WysiwygController, detectUnsupportedMarkdown, parseMarkdown, serializeMarkdown };